@@ -14,6 +14,7 @@
 	"files": [
 		"dist/"
 	],
+	"type": "module",
 	"repository": {
 		"type": "git",
 		"url": "git://github.com/mmgoodnow/cross-seed.git"
@@ -45,7 +46,7 @@
 		"lint-staged": "^12.1.4",
 		"np": "*",
 		"prettier": "^2.0.5",
-		"typescript": "^4.2.3"
+		"typescript": "^4.6.0-dev.20211227"
 	},
 	"prettier": {
 		"tabWidth": 4,
